+++ include/bitcoin/bitcoin/message/get_blocks.hpp
@@ -69,7 +69,7 @@ public:
     void to_data(uint32_t version, writer& sink) const;
     bool is_valid() const;
     void reset();
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     // This class is move assignable but not copy assignable.
     get_blocks& operator=(get_blocks&& other);
