+++ include/bitcoin/bitcoin/message/headers.hpp
@@ -71,7 +71,7 @@ public:
     void to_data(uint32_t version, writer& sink) const;
     bool is_valid() const;
     void reset();
-    uint64_t serialized_size(uint32_t version) const;
+    size_t serialized_size(uint32_t version) const;
 
     // This class is move assignable but not copy assignable.
     headers& operator=(headers&& other);
