+++ src/chain/output.cpp
@@ -217,7 +217,7 @@ std::string output::to_string(uint32_t flags) const
 // Size.
 //-----------------------------------------------------------------------------
 
-uint64_t output::serialized_size(bool wire) const
+size_t output::serialized_size(bool wire) const
 {
     // validation.spender_height is size_t stored as uint32_t.
     return (wire ? 0 : sizeof(uint32_t)) + sizeof(value_) +
